Move dummy device test into a separate script
